requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.1

# Machine learning / AI
torch>=2.0.0
transformers>=4.40.0

# Image processing
pillow>=10.0.0
opencv-python-headless>=4.8.0

# Utilities
numpy>=1.26.0
scikit-learn>=1.5.0
